requirements.txt
pytest
requests
pydantic
allure-pytest
jsonschema
python-dotenv